# railway-backend/openapi.yaml
openapi: 3.0.3
info:
  title: All For Pets — Municipal Pet Registry API
  version: 1.0.0
  description: |
    REST API powering the AFP web + Expo mobile app.

    ### Authentication
    - `POST /api/auth/login` returns `token` (JWT access, 15 min) and
      `refreshToken` (opaque, device-bound, 1–30 days).
    - Send `Authorization: Bearer <token>` on every protected call.
    - On 401 → call `POST /api/auth/refresh` with the refresh token; a **new**
      refresh token is returned (rotation) and the old one is revoked.
    - Mobile clients should send `X-Device-Label` and `X-Expo-Push-Token`
      headers on login/register/refresh so we can list / revoke sessions and
      push notifications.

    ### Caching
    - Public `/api/geo/*` endpoints return `ETag` + `Cache-Control`.
      Send `If-None-Match` to get a 304 with an empty body.

    ### File uploads (mobile)
    - Prefer the `/upload-photo/sign` + `/upload-photo/confirm` flow so the
      binary goes directly to S3/Azure/Cloudinary without touching the API.
servers:
  - url: https://afp.up.railway.app
    description: Production (Railway)
  - url: http://localhost:3000
    description: Local development (Node)
...
security:
  - bearerAuth: []
paths:
...
  /api/geo/cities:
    get:
      tags: [Geo]
      security: []
      summary: Active cities (cached, ETag)
      parameters:
        - { in: header, name: If-None-Match, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/City" } }
        "304": { description: Not modified — use cached copy }
  /api/geo/nigams:
    get:
      tags: [Geo]
      security: []
      parameters:
        - { in: query, name: cityId, schema: { type: integer } }
        - { in: header, name: If-None-Match, schema: { type: string } }
      responses:
        "200":
          description: OK
          content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Nigam" } } } }
        "304": { description: Not modified }
  /api/geo/zones:
    get:
      tags: [Geo]
      security: []
      parameters:
        - { in: query, name: nigamId, schema: { type: integer } }
        - { in: header, name: If-None-Match, schema: { type: string } }
      responses:
        "200":
          description: OK
          content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Zone" } } } }
        "304": { description: Not modified }
  /api/geo/wards:
    get:
      tags: [Geo]
      security: []
      parameters:
        - { in: query, name: zoneId,  schema: { type: integer } }
        - { in: query, name: nigamId, schema: { type: integer } }
        - { in: header, name: If-None-Match, schema: { type: string } }
      responses:
        "200":
          description: OK
          content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Ward" } } } }
        "304": { description: Not modified }

  # ───────────── Pets ─────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    City:
      type: object
      properties:
        id:    { type: integer }
        name:  { type: string }
        state: { type: string, nullable: true }
    Nigam:
      type: object
      properties:
        id:               { type: integer }
        name:             { type: string }
        city_id:          { type: integer }
        registration_fee: { type: number }
        renewal_fee:      { type: number }
        transfer_fee:     { type: number }
    Zone:
      type: object
      properties:
        id:       { type: integer }
        name:     { type: string }
        nigam_id: { type: integer }
    Ward:
      type: object
      properties:
        id:          { type: integer }
        ward_number: { type: string }
        zone_id:     { type: integer, nullable: true }
        nigam_id:    { type: integer, nullable: true }
...